Rewrite event handlers and simplify logic
These changes made it very easy to reorder the laps and to synchronize
the laps displayed in HTML and in the backend.
Also now the start and stop date are included in the state JSON.
With simplification is meant that unncessary variables that could be
derived from others were deleted and that many index logic parts were
removed if the index was not necessary but could be derived from
elsewhere.

Closes #5